+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Added IAM User assertions: `has.name`, `was.created_at`, `belongs_to.group`
 - Added IAM Role assertions: `has.name`, `has.description`, `was.created_at`, `was.last_updated_at`, `should_be.attached_to`
+- Added IAM Group assertions: `has.name`, `was.created_at`, `should.contain`, `has.policy`
 
 
 0.0.13 (2021-04-10)