+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - `s3.Bucket.should.contain()` now accepts `s3.Object` as an alternative to a `str` key
 - assertions now `assert` themselves rather than returning a `bool`
-
+- Added better `AssertionError` messages
 
 0.0.13 (2021-04-10)
 -------------------