+++ CHANGELOG.md
@@ -7,12 +7,14 @@
 - Added IAM Policy assertions: `has.name`, `has.description`, `was.created_at`, `was.last_updated_at`, `should_be.attached_to`
 - Added IAM Group assertions: `has.name`, `was.created_at`, `should.contain`, `has.policy`
 - Added IAM Role assertions: `has.name`, `has.description`, `was.created_at`, `was.last_used_on`, `uses.policy`
+- Added SNS Topic assertions: `should.receive`
 
 ### :pencil2: Changed
 
 - `s3.Bucket.should.contain()` now accepts `s3.Object` as an alternative to a `str` key
 - assertions now `assert` themselves rather than returning a `bool`
 - Added better `AssertionError` messages
+
 
 0.0.13 (2021-04-10)
 -------------------